==> 1 - Scraping.ipynb <==
# **1 - Scraping**

## Importing Libraries

from bs4 import BeautifulSoup
import requests
import pandas as pd
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

## Scraping Fighters
**Example Fighters**: http://www.ufcstats.com/statistics/fighters?char=a&page=all

**Example Fighter**: http://www.ufcstats.com/fighter-details/93fe7332d16c6ad9

### Fetching

fighters_soups = []
# Fighters' Pages Sorted by Last Initial
for letter in 'abcdefghijklmnopqrstuvwxyz':
    fighters_url = f'http://www.ufcstats.com/statistics/fighters?char={letter}&page=all'
    response = requests.get(fighters_url)
    response.raise_for_status
    page = response.text
    soup = BeautifulSoup(page, 'html.parser')
    # print(soup.prettify())
    fighters_soups.append(soup)

### Scraping

# Create DataFrame
fighters = pd.DataFrame(columns = ['Name', 'Record', 'Height', 'Weight', 'Reach', 'Stance', 'Birthday', 'SLPM', 'SACC', 'SAPM', 'SD', 'TDAVG', 'TDACC', 'TDD', 'SAVG' ])

# Loop Through Fighters' Pages (Last Initial)
for soup in fighters_soups:
    fighters_table = soup.find('table')
    fighters_hyperlinks = fighters_table.find_all('a')
    # Loop Through Fighters w/ Same Last Initial
    for fighters_hyperlink in fighters_hyperlinks[::3]:
        fighters_url = fighters_hyperlink['href']
        fighters_response = requests.get(fighters_url)
        fighters_response.raise_for_status()
        fighters_page = fighters_response.text
        fighters_soup = BeautifulSoup(fighters_page, 'html.parser')
    
        # Name
        name  = fighters_soup.find('span', class_ = 'b-content__title-highlight').text.strip()
        # Record
        record = fighters_soup.find('span', class_ = 'b-content__title-record').text.strip().replace('Record: ', '')
        # Height
        height = fighters_soup.find('i', string = lambda text: text and 'height' in text.lower()).next_sibling.strip()
        # Weight
        weight = fighters_soup.find('i', string = lambda text: text and 'weight' in text.lower()).next_sibling.strip()
        # Reach
        reach = fighters_soup.find('i', string = lambda text: text and 'reach' in text.lower()).next_sibling.strip()
        # Stance
        stance = fighters_soup.find('i', string = lambda text: text and 'stance' in text.lower()).next_sibling.strip()
        # Birthday
        birthday = fighters_soup.find('i', string = lambda text: text and 'dob' in text.lower()).next_sibling.strip()
        # SLPM
        slpm = fighters_soup.find('i', string = lambda text: text and 'slpm' in text.lower()).next_sibling.strip()
        # SACC
        sacc = fighters_soup.find('i', string = lambda text: text and 'str. acc' in text.lower()).next_sibling.strip()
        # SAPM
        sapm = fighters_soup.find('i', string = lambda text: text and 'sapm' in text.lower()).next_sibling.strip()
        # SD
        sd = fighters_soup.find('i', string = lambda text: text and 'str. def' in text.lower()).next_sibling.strip()
        # TDAVG
        tdavg = fighters_soup.find('i', string = lambda text: text and 'td avg' in text.lower()).next_sibling.strip()
        # TDACC
        tdacc = fighters_soup.find('i', string = lambda text: text and 'td acc' in text.lower()).next_sibling.strip()
        # TDD
        tdd = fighters_soup.find('i', string = lambda text: text and 'td def' in text.lower()).next_sibling.strip()
        # SAVG
        savg = fighters_soup.find('i', string = lambda text: text and 'sub. avg' in text.lower()).next_sibling.strip()

        # Add to DataFrame
        fighters.loc[len(fighters)] = [name, record, height, weight, reach, stance, birthday, slpm, sacc, sapm, sd, tdavg, tdacc, tdd, savg]

display(fighters.head())
display(fighters.tail())

### Exporting

fighters.to_csv('fighters.csv', index = False)

## Scraping Fights

**Example Events**: http://www.ufcstats.com/statistics/events/completed?page=all

**Example Fights**: http://www.ufcstats.com/statistics/events/completed

**Example Fight**: http://www.ufcstats.com/fight-details/15805ae1eea3343e

### Fetching

fights_url = 'http://www.ufcstats.com/statistics/events/completed?page=all'
response = requests.get(fights_url)
response.raise_for_status
page = response.text
soup = BeautifulSoup(page, 'html.parser')
# print(soup.prettify())

### Scraping

# Create DataFrame
fights = pd.DataFrame(columns = ['A', 'B', 'Winner', 'Event', 'Date', 'Location', 'Division', 'Method', 'Round', 'Time', 'Format', 'Referee',
'A_KD', 'A_SS', 'A_SSP', 'A_TS', 'A_TD', 'A_TDP', 'A_SA', 'A_REV', 'A_CTRL', 'A_H', 'A_B', 'A_L', 'A_D', 'A_C', 'A_G',
'B_KD', 'B_SS', 'B_SSP', 'B_TS', 'B_TD', 'B_TDP', 'B_SA', 'B_REV', 'B_CTRL', 'B_H', 'B_B', 'B_L', 'B_D', 'B_C', 'B_G'
])

events_table = soup.find('table')
events_table_rows = events_table.find_all('tr', class_ = 'b-statistics__table-row')

# Loop Through Events
for events_table_row in events_table_rows[2:]:
    event_name = events_table_row.find('a').text.strip() # Event
    event_date = events_table_row.find('span').text.strip() # Date
    event_location = events_table_row.find('td', class_ = 'b-statistics__table-col b-statistics__table-col_style_big-top-padding').text.strip() # Location

    event_link = events_table_row.find('a')['href']
    event_page = requests.get(event_link).text # Click Event Hyperlink (e.g., http://www.ufcstats.com/event-details/be8ad887e4d674b0)
    event_soup = BeautifulSoup(event_page, 'html')

    # Loop Through Fights in Event
    fight_hyperlinks = event_soup.find_all('tr', class_ = 'b-fight-details__table-row b-fight-details__table-row__hover js-fight-details-click')
    for fight_hyperlink in fight_hyperlinks: 
        fight_hyperlink = fight_hyperlink['data-link'] # Get Fight Hyperlink
        fights_page = requests.get(fight_hyperlink).text # Click Fight Hyperink (e.g., http://www.ufcstats.com/fight-details/15805ae1eea3343e)
        fight_soup = BeautifulSoup(fights_page, 'html')

        # Header
        a = fight_soup.find_all('a', class_ = 'b-link b-fight-details__person-link')[0].text.strip() # A
        b = fight_soup.find_all('a', class_ = 'b-link b-fight-details__person-link')[1].text.strip() # B

        # Winner
        if fight_soup.find('div', class_ = 'b-fight-details__person').find().text.strip() == 'W':
            winner = a
        elif fight_soup.find('div', class_ = 'b-fight-details__person').find().text.strip() == 'D':
            winner = 'Draw'
        else:
            winner = b

        division = fight_soup.find('i', class_ = 'b-fight-details__fight-title').text.strip() # Division
        method = fight_soup.find('i', string = lambda text: text and 'method' in text.lower()).find_next('i').text.strip() # Method
        round = fight_soup.find('i', string = lambda text: text and 'round' in text.lower()).next_sibling.strip() # Round
        time = fight_soup.find('i', string = lambda text: text and 'time' in text.lower()).next_sibling.strip() # Time
        format = fight_soup.find('i', string = lambda text: text and 'format' in text.lower()).next_sibling.strip() # Format
        referee = fight_soup.find('i', string = lambda text: text and 'referee' in text.lower()).find_next('span').text.strip() # Referee

        # Skips Fights w/o Table 1 and 2
        if fight_soup.find('section', class_ = 'b-fight-details__section js-fight-section').text.strip() == 'Round-by-round stats not currently available.':
            fights.loc[len(fights)] = [a, b, winner, event_name, event_date, event_location, division, method, round, time, format, referee,
        None, None, None, None, None, None, None, None, None, None, None, None, None, None, None,
        None, None, None, None, None, None, None, None, None, None, None, None, None, None, None
        ]
            
        else:
            # Table 1
            fight_body_1_soup = fight_soup.find_all('tbody', class_ = 'b-fight-details__table-body')[0]
            table_1_columns = fight_body_1_soup.find_all('p', class_ = 'b-fight-details__table-text')
            del table_1_columns[0:2] # Delete Irrelevant Columns
            a_kd = table_1_columns[0].text.strip()
            b_kd = table_1_columns[1].text.strip()
            a_ss = table_1_columns[2].text.strip()
            b_ss = table_1_columns[3].text.strip()
            a_ssp = table_1_columns[4].text.strip()
            b_ssp = table_1_columns[5].text.strip()
            a_ts = table_1_columns[6].text.strip()
            b_ts = table_1_columns[7].text.strip()
            a_td = table_1_columns[8].text.strip()
            b_td = table_1_columns[9].text.strip()
            a_tdp = table_1_columns[10].text.strip()
            b_tdp = table_1_columns[11].text.strip()
            a_sa = table_1_columns[12].text.strip()
            b_sa = table_1_columns[13].text.strip()
            a_rev = table_1_columns[14].text.strip()
            b_rev = table_1_columns[15].text.strip()
            a_ctrl = table_1_columns[16].text.strip()
            b_ctrl = table_1_columns[17].text.strip()

            # Table 2
            fight_body_2_soup = fight_soup.find_all('tbody', class_ = 'b-fight-details__table-body')[2]
            table_2_columns = fight_body_2_soup.find_all('p', class_ = 'b-fight-details__table-text')
            del table_2_columns[0:6] # Delete Irrelevant Columns
            a_h = table_2_columns[0].text.strip()
            b_h = table_2_columns[1].text.strip()
            a_b = table_2_columns[2].text.strip()
            b_b = table_2_columns[3].text.strip()
            a_l = table_2_columns[4].text.strip()
            b_l = table_2_columns[5].text.strip()
            a_d = table_2_columns[6].text.strip()
            b_d = table_2_columns[7].text.strip()
            a_c = table_2_columns[8].text.strip()
            b_c = table_2_columns[9].text.strip()
            a_g = table_2_columns[10].text.strip()
            b_g = table_2_columns[11].text.strip()

            # Add to DataFrame
            fights.loc[len(fights)] = [a, b, winner, event_name, event_date, event_location, division, method, round, time, format, referee,
            a_kd, a_ss, a_ssp, a_ts, a_td, a_tdp, a_sa, a_rev, a_ctrl, a_h, a_b, a_l, a_d, a_c, a_g,
            b_kd, b_ss, b_ssp, b_ts, b_td, b_tdp, b_sa, b_rev, b_ctrl, b_h, b_b, b_l, b_d, b_c, b_g
            ]

print('Head:')
display(fights.head())
print('Tail:')
display(fights.tail())

### Exporting

fights.to_csv('fights.csv', index = False)
