==> 2 - Loading.ipynb <==
# **2 - Loading**

## Importing Libraries

import numpy as np
import pandas as pd
import pickle
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

## Reading CSV

fighters = pd.read_csv('./fighters.csv', sep = ',', header = 0)
fights = pd.read_csv('fights.csv', sep = ',', header = 0)

## Overview

print('Fighters Overview')
print('************************************')
print('Shape:', fighters.shape)
print('------------------------------------')
print('Columns:', fighters.columns.values)
print('------------------------------------')
print('Head:')
display(fighters.head())
print('Tail:')
display(fighters.tail())

print('Fights Overview')
print('************************************')
print('Shape:', fights.shape)
print('------------------------------------')
print('Columns:', fights.columns.values)
print('------------------------------------')
print('Head:')
display(fights.head())
print('Tail:')
display(fights.tail())

## Description

print('Fighters Description')
print('************************************')
display(fighters.info())
print('------------------------------------------------------------------------')
display(fighters.describe(include = ['object']))
display(fighters.describe(include = ['float64']))

print('Fights Description')
print('************************************')
display(fights.info())
print('------------------------------------------------------------------------')
display(fights.describe(include = ['object']))
display(fights.describe(include = ['int64']))

## Missing  Values


# Replace "-", "--", "---", and None with np.nan
fighters = fighters.replace('-', np.nan)
fighters = fighters.replace('--', np.nan)
fighters = fighters.replace('---', np.nan)
fighters = fighters.fillna(np.nan)
fights = fights.replace('-', np.nan)
fights = fights.replace('--', np.nan)
fights = fights.replace('---', np.nan)
fights = fights.fillna(np.nan)

print('Fighters Missing Values')
print('************************************')
for column in fighters.columns:
    print(column, 'Missing Values:', fighters[column].isnull().sum())
    
print('\nFights Missing Values')
print('************************************')
for column in fights.columns:
    print(column, 'Missing Values:', fights[column].isnull().sum())

## Duplicate Values

print(f'Fighters Duplicate Values: {fighters.duplicated().sum()}')
print(f'Fights Duplicate Values: {fights.duplicated().sum()}')

## Unique Values

print('Fighters Unique Values')
print('************************************')
for column in fighters.columns:
    print(column, 'Unique Values:', fighters[column].nunique())
    
print('')
print('Fights Unique Values')
print('************************************')
for column in fights.columns:
    print(column, 'Unique Values:', fights[column].nunique())

## Exporting Pickles

with open('fighters.pkl', 'wb') as f:
    pickle.dump(fighters, f)
with open('fights.pkl', 'wb') as f:
    pickle.dump(fights, f)
