==> 3 - Preprocessing.ipynb <==
# **3 - Data Preprocessing**

## Importing Libraries

import numpy as np
import pandas as pd
import pickle
pd.set_option('display.max_columns', None)

## Importing Pickles

with open('fighters.pkl', 'rb') as f:
    fighters = pickle.load(f)
with open('fights.pkl', 'rb') as f:
    fights = pickle.load(f)

## Dropping Columns

# Dropping Post-Fight Columns
fights = fights[['A', 'B', 'Winner', 'Event', 'Date', 'Location', 'Division', 'Format', 'Referee']]

## Dropping Rows

# Dropping Draws
print(f'Before # of Draws: {fights[fights['Winner'] == 'Draw'].shape[0]}')
fights = fights[fights['Winner'] != 'Draw']
print(f'After # of Draws: {fights[fights["Winner"] == 'Draw'].shape[0]}')

## Shuffling

# B = Scraped Default Winner. Shuffle A and B.
print('Before:')
display(fights.head(5))

mask = np.random.rand(len(fights)) < 0.5
fights.loc[mask, ['A', 'B']] = fights.loc[mask, ['B', 'A']].values

print('After:')
display(fights.head(5))

## Merging DataFrames

# Before
display(fights.iloc[[0], :])
display(fighters.loc[(fighters['Name'] == "Sean O'Malley") | (fighters['Name'] == 'Merab Dvalishvili')])

df = pd.merge(fights, fighters, left_on = 'A', right_on = 'Name', how = 'left')
df = pd.merge(df, fighters, left_on = 'B', right_on = 'Name', how = 'left')

# Renaming Columns
df.columns = ['A', 'B', 'Winner', 'Event', 'Date', 'Location', 'Division', 'Format', 'Referee', 
              'A_Name', 'A_Record', 'A_Height', 'A_Weight', 'A_Reach', 'A_Stance', 'A_Birthday', 'A_SLPM', 'A_SACC', 'A_SAPM', 'A_SD', 'A_TDAVG', 'A_TDACC', 'A_TDD', 'A_SAVG',
              'B_Name', 'B_Record', 'B_Height', 'B_Weight', 'B_Reach', 'B_Stance', 'B_Birthday', 'B_SLPM', 'B_SACC', 'B_SAPM', 'B_SD', 'B_TDAVG', 'B_TDACC', 'B_TDD', 'B_SAVG']

# After
display(df.iloc[[0], :])

## Modifying Columns

### Converting Height, Weight, and Reach

print(f'Before:')
display(df[['A_Height', 'A_Weight', 'A_Reach', 'B_Height', 'B_Weight', 'B_Reach']].head())

# Convert from FT IN to CM
def convert_height(height):
    if pd.isna(height):
        return np.nan
    feet, inches = height.split(" ")
    feet = float(feet[0])
    inches = float(inches[0])
    return (feet * 30.48) + (inches * 2.54)

# Convert from LB to KG
def convert_weight(weight):
    if pd.isna(weight):
        return np.nan
    weight = weight.split(" ")
    weight = float(weight[0])
    return weight * 0.4536

# Convert from IN to CM
def convert_reach(reach):
    if pd.isna(reach):
        return np.nan
    reach = float(reach[0:2])
    return reach * 2.54

df['A_Height'] = df['A_Height'].apply(convert_height)
df['A_Weight'] = df['A_Weight'].apply(convert_weight)
df['A_Reach'] = df['A_Reach'].apply(convert_reach)
df['B_Height'] = df['B_Height'].apply(convert_height)
df['B_Weight'] = df['B_Weight'].apply(convert_weight)
df['B_Reach'] = df['B_Reach'].apply(convert_reach)

print(f'After:')
display(df[['A_Height', 'A_Weight', 'A_Reach', 'B_Height', 'B_Weight', 'B_Reach']].head())

### Calculating Age

# Convert to DateTime
df['Date'] = pd.to_datetime(df['Date'])
df['A_Birthday'] = pd.to_datetime(df['A_Birthday'])
df['B_Birthday'] = pd.to_datetime(df['B_Birthday'])
# Age in Days
df['A_Age'] = (df['Date'] - df['A_Birthday']).dt.days
df['B_Age'] = (df['Date'] - df['B_Birthday']).dt.days

display(df[['Date', 'A_Birthday', 'B_Birthday', 'A_Age', 'B_Age']].head())

### Binarize Winner

print('Before:')
display(df[['A', 'B', 'Winner']].head())

def binarize_winner(row):
    if row['Winner'] == row['A']:
        return 1
    else:
        return 0
df['Winner'] = df.apply(binarize_winner, axis = 1)

print('After:')
display(df[['A', 'B', 'Winner']].head())

## Handle Missing Values

# Imputing Reach = Height

print('Before')
print('------------------------------------')
for column in df.columns:
    print(column, 'Missing Values:', df[column].isnull().sum())

df['A_Reach'] = df['A_Reach'].fillna(df['A_Height'])
df['B_Reach'] = df['B_Reach'].fillna(df['B_Height'])

print('')
print('After')
print('------------------------------------')
for column in df.columns:
    print(column, 'Missing Values:', df[column].isnull().sum())

## Reorder Columns

print('Before:')
display(df.head(5))

df['Fight'] = df['A'] + ' vs. ' + df['B']
df = df[['Event', 'Fight', 'A', 'B', 'Winner', 'Date', 'Location', 'Division', 'Format', 'Referee', 
         'A_Record', 'A_Height', 'A_Weight', 'A_Reach', 'A_Stance', 'A_Birthday', 'A_Age', 'A_SLPM', 'A_SACC', 'A_SAPM', 'A_SD', 'A_TDAVG', 'A_TDACC', 'A_TDD', 'A_SAVG',
         'B_Record', 'B_Height', 'B_Weight', 'B_Reach', 'B_Stance', 'B_Birthday', 'B_Age', 'B_SLPM', 'B_SACC', 'B_SAPM', 'B_SD', 'B_TDAVG', 'B_TDACC', 'B_TDD', 'B_SAVG']]

print('')
print('After:')
display(df.head(5))

## Changing Data Types

# Before
print('Before:')
print(df.info())

# Converting Percentages to Floats
cols = ['A_SACC', 'A_SD', 'A_TDACC', 'A_TDD', 'B_SACC', 'B_SD', 'B_TDACC', 'B_TDD']

for col in cols:
    df[col] = df[col].astype(str).str.rstrip('%').astype('float') / 100.0

# Converting into Floats, Objects (Strings), Datetime, and Categories
dtype_dict = {
    'Event': 'object', 'Fight': 'object', 'A': 'object', 'B': 'object', 'Winner': 'int64', 'Date': 'datetime64[ns]', 'Location': 'object', 'Division': 'category', 'Format': 'category', 'Referee': 'object',
    'A_Record': 'object', 'A_Height': 'float64', 'A_Weight': 'float64', 'A_Reach': 'float64', 'A_Stance': 'category', 'A_Birthday': 'datetime64[ns]', 'A_Age': 'float64', 'A_SLPM': 'float64', 'A_SACC': 'float64', 'A_SAPM': 'float64', 'A_SD': 'float64', 'A_TDAVG': 'float64', 'A_TDACC': 'float64', 'A_TDD': 'float64', 'A_SAVG': 'float64',
    'B_Record': 'object', 'B_Height': 'float64', 'B_Weight': 'float64', 'B_Reach': 'float64', 'B_Stance': 'category', 'B_Birthday': 'datetime64[ns]', 'B_Age': 'float64', 'B_SLPM': 'float64', 'B_SACC': 'float64', 'B_SAPM': 'float64', 'B_SD': 'float64', 'B_TDAVG': 'float64', 'B_TDACC': 'float64', 'B_TDD': 'float64', 'B_SAVG': 'float64'
}

df = df.astype(dtype_dict)

# After
print('')
print('After:')
print(df.info())

## Exporting Pickle

with open('df.pkl', 'wb') as f:
    pickle.dump(df, f)
